Split commands into separate files
