+++ src/commands/index.ts
@@ -1,0 +1,5 @@
+export { migrate } from "./migrate";
+export { watch } from "./watch";
+export { reset } from "./reset";
+export { commit } from "./commit";
+export { status } from "./status";
